fastapi/requirements.txt
llama-stack
ipykernel==6.29.5
langchain==0.3.7
langchain-together==0.2.0
langchain-community==0.3.5
pypdf
faiss-cpu
sentence-transformers
python-dotenv==1.0.1
fastapi==0.115.0
uvicorn==0.30.1
pydantic==2.8.2
together==1.3.3
